+++ cacher.js
@@ -1,22 +1,22 @@
-var queue = require('./queue');
+/**
+ * Worker to handle counting and caching tasks performed when a match is inserted or parsed.
+ * All operations in this worker should deal with ephemeral data (can be reconstructed from persistent data stores)
+ **/
+var queue = require('./store/queue');
 var cQueue = queue.getQueue('cache');
-var utility = require('./utility');
-var redis = require('./redis');
+var utility = require('./util/utility');
+var redis = require('./store/redis');
 var moment = require('moment');
-var benchmarks = require('./benchmarks');
+var benchmarks = require('./compute/benchmarks');
 var async = require('async');
 var constants = require('./constants');
 var config = require('./config');
-var getMatchRating = require('./getMatchRating');
+var getMatchRating = require('./util/getMatchRating');
 cQueue.process(10, processCache);
 cQueue.on('completed', function(job)
 {
     job.remove();
 });
-/**
- * Handles counting and caching tasks to be performed when a match is inserted or parsed.
- * All operations in this processor should deal with ephemeral data (can be reconstructed from persistent data stores)
- **/
 function processCache(job, cb)
 {
     var match = job.data.payload;
