+++ monitor.js
@@ -1,8 +1,11 @@
-var redis = require('./redis');
-var db = require('./db');
+/**
+ * Worker that monitors health metrics and saves results
+ **/
+var redis = require('./store/redis');
+var db = require('./store/db');
+var utility = require('./util/utility');
 var config = require('./config');
 var request = require('request');
-var utility = require('./utility');
 var api_key = config.STEAM_API_KEY.split(',')[0];
 var health = {
     random_match: function random_match(cb)
