+++ profiler.js
@@ -1,10 +1,13 @@
-var queries = require('./queries');
+/**
+ * Worker to fetch updated player profiles
+ **/
+var constants = require('./constants');
+var queries = require('./store/queries');
+var db = require('./store/db');
+var utility = require('./util/utility');
 var insertPlayer = queries.insertPlayer;
-var utility = require('./utility');
 var getData = utility.getData;
 var async = require('async');
-var db = require('./db');
-var constants = require('./constants');
 var max;
 start();
