+++ proxy.js
@@ -1,3 +1,13 @@
+/**
+ * Worker proxying requests to the Steam API.
+ **/
+//mirrors steam api
+var config = require('./config');
+var httpProxy = require('http-proxy');
+httpProxy.createProxyServer({
+    target: 'http://api.steampowered.com',
+    changeOrigin: true
+}).listen(config.PORT || config.OPENSHIFT_NODEJS_PORT || config.PROXY_PORT, config.OPENSHIFT_NODEJS_IP);
 /*
 //general purpose proxy
 var express = require('express');
@@ -9,10 +19,3 @@ app.use(function(req, res) {
 });
 app.listen(config.PORT);
 */
-//mirrors steam api
-var config = require('./config');
-var httpProxy = require('http-proxy');
-httpProxy.createProxyServer({
-    target: 'http://api.steampowered.com',
-    changeOrigin: true
-}).listen(config.PORT || config.OPENSHIFT_NODEJS_PORT || config.PROXY_PORT, config.OPENSHIFT_NODEJS_IP);
