+++ store/buildPlayer.js
@@ -1,10 +1,10 @@
 var async = require('async');
-var constants = require('./constants.js');
-var queries = require("./queries");
-var utility = require('./utility');
-var config = require('./config');
-var playerCache = require('./playerCache');
-var aggregator = require('./aggregator');
+var constants = require('../constants.js');
+var queries = require("../store/queries");
+var utility = require('../util/utility');
+var config = require('../config');
+var playerCache = require('../store/playerCache');
+var aggregator = require('../compute/aggregator');
 var getPlayerMatches = queries.getPlayerMatches;
 var getPlayer = queries.getPlayer;
 var getPlayerRankings = queries.getPlayerRankings;
