+++ store/cassandra.js
@@ -1,5 +1,8 @@
+/**
+ * Interface to Cassandra client
+ **/
 var cass = require('cassandra-driver');
-var config = require('./config');
+var config = require('../config');
 var url = require('url');
 var u = url.parse(config.CASSANDRA_URL);
 console.error('connecting %s', config.CASSANDRA_URL);
