+++ store/queue.js
@@ -1,8 +1,11 @@
+/**
+ * Provides methods for working with the job queue
+ **/
+var generateJob = require('../util/utility').generateJob;
+var config = require('../config');
 var bull = require('bull');
-var config = require('./config');
 var url = require('url');
 var async = require('async');
-var generateJob = require('./utility').generateJob;
 // parse the url
 var conn_info = url.parse(config.REDIS_URL, true /* parse query string */ );
 if (conn_info.protocol !== 'redis:')
