+++ webpack.config.js
@@ -1,3 +1,6 @@
+/**
+ * Webpack configuration file
+ **/
 var webpack = require('webpack');
 var config = require('./config');
 module.exports = {
